+++ src/Exception/JobTimeoutException.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace KinetiStack\Sdk\Exception;
 
-use KinetiStack\Sdk\Dto\BatchJobDto;
+use KinetiStack\Sdk\Dto\JobDto;
 
-class BatchJobTimeoutException extends KinetiException
+class JobTimeoutException extends KinetiException
 {
     public function __construct(
         string $message,
-        public readonly BatchJobDto $latestJob,
+        public readonly JobDto $latestJob,
         ?\Throwable $previous = null
     ) {
         parent::__construct($message, 0, $previous);